+++ test/full_flow/error_recovery/sql_statements/sql_recovery_parser.ts
@@ -56,15 +56,15 @@ export class DDLExampleRecoveryParser extends Parser {
     // with the error recovery re-sync behavior.
     // note that when one parsing rule calls another (via SUBRULE) the invoked rule is the one defined here,
     // without the "parse" prefix.
-    public ddl = this.RULE("ddl", this.parseDdl, INVALID(INVALID_DDL))
+    public ddl = this.RULE("ddl", this.parseDdl, {recoveryValueFunc: INVALID(INVALID_DDL)})
     // DOCS: a specific return type has been provided in case of re-sync recovery.
-    public createStmt = this.RULE("createStmt", this.parseCreateStmt, INVALID(INVALID_CREATE_STMT))
-    public insertStmt = this.RULE("insertStmt", this.parseInsertStmt, INVALID(INVALID_INSERT_STMT))
-    public deleteStmt = this.RULE("deleteStmt", this.parseDeleteStmt, INVALID(INVALID_DELETE_STMT))
-    public qualifiedName = this.RULE("qualifiedName", this.parseQualifiedName, INVALID(INVALID_QUALIFIED_NAME))
-    public recordValue = this.RULE("recordValue", this.parseRecordValue, INVALID())
+    public createStmt = this.RULE("createStmt", this.parseCreateStmt, {recoveryValueFunc: INVALID(INVALID_CREATE_STMT)})
+    public insertStmt = this.RULE("insertStmt", this.parseInsertStmt, {recoveryValueFunc: INVALID(INVALID_INSERT_STMT)})
+    public deleteStmt = this.RULE("deleteStmt", this.parseDeleteStmt, {recoveryValueFunc: INVALID(INVALID_DELETE_STMT)})
+    public qualifiedName = this.RULE("qualifiedName", this.parseQualifiedName, {recoveryValueFunc: INVALID(INVALID_QUALIFIED_NAME)})
+    public recordValue = this.RULE("recordValue", this.parseRecordValue, {recoveryValueFunc: INVALID()})
     // DOCS: A Parsing rule may also be private and not part of the public API
-    private value = this.RULE("value", this.parseValue, INVALID())
+    private value = this.RULE("value", this.parseValue, {recoveryValueFunc: INVALID()})
 
     // DOCS: note how all the parsing rules in this example return a ParseTree, we require some output from the parser
     // to demonstrate the error recovery mechanisms. otherwise it is harder to prove we have indeed recovered.